some changes

+++ src/main/java/com/example/datagaze_task/entity/Country.java
@@ -33,7 +33,7 @@ public class Country {
 
     private String wind_color;
 
-    private Long cloud;
+    private Integer cloud;
 
     private String cloud_color;
 }
